fix: Use Uint8Array instead of ArrayBuffer in Decoder
This PR:

1. Removes the type-wise redundant `instanceof` check from `Decoder.load`
2. Uses `Uint8Array` instead of `ArrayBuffer`

On 1, `instanceof` is not always reliable due the vagaries of different
Javascript environments and embeddings (e.g. reference equality of
ArrayBuffer may not hold). In my case a `Buffer` instance returned false
for `buf instanceof ArrayBuffer. Besides that type-wise the `instanceof`
check is redundant and it's kind of nasty to have `load` silently
actually not do a load.

On 2, seems like there's no reason not to use the `Uint8Array` which is
a bit more standard in general and in this repo, I also believe it has
more reliable inheritance.

Signed-off-by: Silas Davis <silas@inco.network>